+++ neuronlp2/io/conllx_data.py
@@ -154,11 +154,11 @@ def get_batch(data, batch_size):
 
     bucket_length = _buckets[bucket_id]
 
-    wid_inputs = np.empty([batch_size, bucket_length], dtype=np.int32)
-    cid_inputs = np.empty([batch_size, bucket_length, utils.MAX_CHAR_LENGTH], dtype=np.int32)
-    pid_inputs = np.empty([batch_size, bucket_length], dtype=np.int32)
-    hid_inputs = np.empty([batch_size, bucket_length], dtype=np.int32)
-    tid_inputs = np.empty([batch_size, bucket_length], dtype=np.int32)
+    wid_inputs = np.empty([batch_size, bucket_length])
+    cid_inputs = np.empty([batch_size, bucket_length, utils.MAX_CHAR_LENGTH])
+    pid_inputs = np.empty([batch_size, bucket_length])
+    hid_inputs = np.empty([batch_size, bucket_length])
+    tid_inputs = np.empty([batch_size, bucket_length])
 
     masks = np.zeros([batch_size, bucket_length], dtype=np.float32)
 
@@ -201,11 +201,11 @@ def iterate_batch(data, batch_size, shuffle=False):
             continue
 
         bucket_length = _buckets[bucket_id]
-        wid_inputs = np.empty([bucket_size, bucket_length], dtype=np.int32)
-        cid_inputs = np.empty([bucket_size, bucket_length, utils.MAX_CHAR_LENGTH], dtype=np.int32)
-        pid_inputs = np.empty([bucket_size, bucket_length], dtype=np.int32)
-        hid_inputs = np.empty([bucket_size, bucket_length], dtype=np.int32)
-        tid_inputs = np.empty([bucket_size, bucket_length], dtype=np.int32)
+        wid_inputs = np.empty([bucket_size, bucket_length])
+        cid_inputs = np.empty([bucket_size, bucket_length, utils.MAX_CHAR_LENGTH])
+        pid_inputs = np.empty([bucket_size, bucket_length])
+        hid_inputs = np.empty([bucket_size, bucket_length])
+        tid_inputs = np.empty([bucket_size, bucket_length])
 
         masks = np.zeros([bucket_size, bucket_length], dtype=np.float32)
